fixed failing query

+++ api/src/queries/getGameStatistics.sql
@@ -1,6 +1,6 @@
 SELECT
-    ROUND(AVG(EXTRACT(EPOCH FROM ("end" - "start"))))::INT AS avg_seconds,
-    ROUND(SUM(EXTRACT(EPOCH FROM ("end" - "start")) / 60.0))::INT AS total_minutes,
+    ROUND(AVG(EXTRACT(EPOCH FROM ("end" - "start"))))::BIGINT AS avg_seconds,
+    ROUND(SUM(EXTRACT(EPOCH FROM ("end" - "start")) / 60.0))::BIGINT AS total_minutes,
     ROUND(
         SUM(
             CASE WHEN DATE_PART('year', "end") = DATE_PART('year', CURRENT_DATE)
@@ -8,6 +8,6 @@ SELECT
                  ELSE 0
             END
         )
-    )::INT AS total_minutes_this_year,
-    COUNT(*)::INT AS total_sessions
+    )::BIGINT AS total_minutes_this_year,
+    COUNT(*)::BIGINT AS total_sessions
 FROM game_activity;
